+++ main.go
@@ -16,7 +16,7 @@ import (
 
 // Build information (set by ldflags)
 var (
-	version   = "4.1.3"
+	version   = "4.1.4"
 	buildTime = "unknown"
 	gitCommit = "unknown"
 )